Move minion auth into salt.master.AuthFuncs
Auth handling lived inline on ReqServerChannel, intercepted before any
worker dispatch.  With worker pools enabled, that left the auth
implementation in the channel layer even though pool routing already
ran the channel inside an MWorker — and the routing-process fallback
could still execute auth outside any worker.

Lift _auth and its helpers (_clear_signed, session_key, compare_keys,
aes_key) into a new AuthFuncs(TransportMethods) class in salt/master.py
alongside ClearFuncs/AESFuncs.  ReqServerChannel._auth becomes a thin
delegator that builds an AuthFuncs from the channel's state and forwards
the call, preserving the existing test API surface (callers that
monkey-patch attributes on the channel see those changes reflected in
the auth handler).

Default DEFAULT_WORKER_POOLS now ships an auth pool (worker_count=1,
commands=[_auth]) plus a default catchall pool (worker_count=4) so the
total worker count matches the historical worker_threads=5 default but
auth is isolated from slow workloads (slow ext_pillar, etc.).  The
single auth worker avoids the documented salt.cache.Cache race on
pending → accepted → denied transitions; operators that have audited
the cache backend for atomic stores can raise this.
